Renamed agave.frames to agave.core
