+++ agave/arp/listen.py
@@ -1,7 +1,7 @@
 import socket, time
 from typing import Tuple
-from agave.frames import ethernet, arp
-from agave.frames.ethernet import MACAddress
+from agave.core import ethernet, arp
+from agave.core.ethernet import MACAddress
 from ipaddress import ip_address, ip_network, IPv4Address
 from .utils import _create_socket, _parse, SOCKET_MAX_READ, ARPReaderLoop, HOST
