+++ agave/arp/rdiscover.py
@@ -1,7 +1,7 @@
 from typing import Iterator, Tuple, Union
 from .utils import _create_socket
 from .resolve import resolve
-from agave.frames.ethernet import MACAddress
+from agave.core.ethernet import MACAddress
 from agave.nic.interfaces import NetworkInterface
 from ipaddress import IPv4Address, IPv4Network, ip_network
