+++ agave/irdp/solicit.py
@@ -8,8 +8,8 @@ Example:
 	python3 -m agave.irdp.solicit
 
 """
-from agave.frames import irdp
-from agave.frames.core import Buffer
+from agave.core import irdp
+from agave.core.buffer import Buffer
 import socket, sys
